+++ monkey/monkey_island/cc/setup/certificate_setup.py
